+++ lib/faker/indian.rb
@@ -5,13 +5,16 @@ require "faker"
 require_relative "indian/version"
 require_relative "indian/data"
 require_relative "indian/name"
+require_relative "indian/languages"
 require_relative "indian/identity"
 require_relative "indian/payment"
 require_relative "indian/vehicle"
 require_relative "indian/address"
 require_relative "indian/phone"
 require_relative "indian/food"
-require_relative "indian/languages"
+require_relative "indian/company"
+require_relative "indian/travel"
+require_relative "indian/profile"
 
 module Faker
   module Indian