+++ lib/faker/indian/payment.rb
@@ -5,15 +5,16 @@ require "faker"
 module Faker
   module Indian
     class Payment
-      _data = Data.load(:payment)
+      data = Data.load(:payment)
 
-      HANDLES = _data.fetch(:handles).freeze
-      BANK_CODES = _data.fetch(:banks).freeze
-      NAMES = _data.fetch(:names).freeze
+      HANDLES = data.fetch(:handles).freeze
+      BANK_CODES = data.fetch(:banks).freeze
+      NAMES = data.fetch(:names).freeze
 
       class << self
-        def upi_id
-          "#{sample_name}#{random.rand(10..99)}@#{sample_handle}"
+        def upi_id(name: nil)
+          handle_name = sanitize_name(name || sample_name)
+          "#{handle_name}#{random.rand(10..99)}@#{sample_handle}"
         end
         alias upi upi_id
 
@@ -29,6 +30,19 @@ module Faker
 
         def account_number
           random.rand(100_000_000_000..999_999_999_999).to_s
+        end
+
+        def demat_account
+          random.rand(100_000_000_000_000..999_999_999_999_999).to_s
+        end
+
+        def bank_details
+          code = sample_bank_code
+          {
+            bank_name: BANK_CODES[code],
+            ifsc: "#{code}0#{random.rand(100_000..999_999)}",
+            account_number: random.rand(100_000_000_000..999_999_999_999).to_s
+          }
         end
 
         private
@@ -47,6 +61,10 @@ module Faker
 
         def sample_bank_code
           BANK_CODES.keys.sample(random: random)
+        end
+
+        def sanitize_name(name)
+          name.to_s.downcase.gsub(/\s+/, "")
         end
       end
     end