+++ spec/faker/indian/address_spec.rb
@@ -7,6 +7,12 @@ RSpec.describe Faker::Indian::Address do
     it "returns a known city" do
       expect(described_class::LOCATIONS.keys).to include(described_class.city)
     end
+
+    it "filters by state" do
+      city = described_class.city(state: "Maharashtra")
+      state = described_class::LOCATIONS.fetch(city).fetch(:state)
+      expect(state).to eq("Maharashtra")
+    end
   end
 
   describe ".state" do
@@ -23,11 +29,35 @@ RSpec.describe Faker::Indian::Address do
     end
   end
 
+  describe ".state_code" do
+    it "returns a known state code" do
+      expect(described_class::STATE_CODES.values).to include(described_class.state_code)
+    end
+  end
+
+  describe ".district" do
+    it "returns a known district" do
+      expect(described_class::DISTRICTS).to include(described_class.district)
+    end
+  end
+
+  describe ".landmark" do
+    it "returns a known landmark" do
+      expect(described_class::LANDMARKS).to include(described_class.landmark)
+    end
+  end
+
+  describe ".line1" do
+    it "returns street line" do
+      expect(described_class.line1).to match(/\A\d+ .+\z/)
+    end
+  end
+
   describe ".full_address" do
     it "includes city, state, and pincode" do
-      address = described_class.full_address
+      address = described_class.full_address(state: "Karnataka")
 
-      expect(address).to match(/\A\d+ .+, .+, .+ - \d{6}\z/)
+      expect(address).to match(/\A\d+ .+, .+, .+, Karnataka - \d{6}\z/)
     end
   end
 end